% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ascii_check.R
\name{ascii_check}
\alias{ascii_check}
\title{ASCII Character Check}
\usage{
ascii_check(DD.dict, DS.data, verbose = TRUE)
}
\arguments{
\item{DD.dict}{Data dictionary.}

\item{DS.data}{Data set.}

\item{verbose}{When TRUE, the function prints the Message out, as well as detailed information on non-ASCII locations.}
}
\value{
Tibble, returned invisibly, containing: (1) Time (Time stamp); (2) Name (Name of the function); (3) Status (Passed/Failed); (4) Message (A copy of the message the function printed out); (5) Information (Column, row, and value of detected non-ASCII characters).
}
\description{
This function scans both the data dictionary and data set for problematic characters that may interfere with dbGaP submission requirements. Specifically, it checks for: (1) Non-ASCII characters (e.g., with accents) and (2) Newline and carriage return characters (e.g., line breaks). It returns a list of any variable names (columns), row numbers, and values where these issues are detected.
}
\examples{
# Passed example
data(ExampleA)
ascii_check(DD.dict.A, DS.data.A) 

# Failed example
data(ExampleT)
ascii_check(DD.dict.T, DS.data.T)
}
