% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_function.R
\name{dat_function}
\alias{dat_function}
\title{Data Utility Function}
\usage{
dat_function(
  DS.dataset,
  DD.dictionary,
  sex.split = FALSE,
  sex.name = NULL,
  DS.dataset.na
)
}
\arguments{
\item{DS.dataset}{Data set.}

\item{DD.dictionary}{Data dictionary.}

\item{sex.split}{When TRUE, split reports by the field named by the sex.name string.}

\item{sex.name}{Character string giving the name of the sex field.}

\item{DS.dataset.na}{Data set with missing values set to NA.}
}
\value{
NULL, called for its side effects.
}
\description{
This function calls eval_function to generate a textual and graphical report of the selected variables.
}
