% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_function_selected.R
\name{dat_function_selected}
\alias{dat_function_selected}
\title{Data Selected Utility Function}
\usage{
dat_function_selected(
  dataset,
  dictionary,
  sex.split = FALSE,
  sex.name = NULL,
  start = 1,
  end = 1,
  dataset.na,
  h.level = 2
)
}
\arguments{
\item{dataset}{Data set.}

\item{dictionary}{Data dictionary.}

\item{sex.split}{When TRUE, split reports by the field named 'Sex'.}

\item{sex.name}{Character string giving the name of the sex field.}

\item{start}{Staring index of the first selected trait.}

\item{end}{Ending index of the last selected trait.}

\item{dataset.na}{Data set with missing values set to NA.}

\item{h.level}{Header level for pandoc function.}
}
\value{
NULL, called for its side effects
}
\description{
This function calls eval_function to generate a textual and graphical report of the selected variables.
}
