% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup_values.R
\name{dup_values}
\alias{dup_values}
\title{Duplicate Values Function}
\usage{
dup_values(DD.dict)
}
\arguments{
\item{DD.dict}{Data dictionary.}
}
\value{
Logical, TRUE if only one VALUES column is detected.
}
\description{
This function checks for duplicate VALUES column names in the data dictionary.
}
