% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_function.R
\name{eval_function}
\alias{eval_function}
\title{Evaluation Utility Function}
\usage{
eval_function(
  dataset,
  dictionary,
  sex.split = FALSE,
  sex.name = NULL,
  dataset.na,
  h.level = 2
)
}
\arguments{
\item{dataset}{Data set.}

\item{dictionary}{Data dictionary.}

\item{sex.split}{When TRUE, split reports by the field named 'Sex'.}

\item{sex.name}{Name of the Sex field.}

\item{dataset.na}{Data set with missing values set to NA.}

\item{h.level}{Header level for pandoc function.}
}
\value{
NULL, called for its side effects.
}
\description{
This function generates a textual and graphical report of the selected variables.
}
