% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleA.R
\docType{data}
\name{ExampleA}
\alias{ExampleA}
\title{ExampleA}
\format{
\code{R data file} that contains two objects:
\describe{
\item{DD.dict.A}{Data dictionary}
\item{DS.data.A}{Data set}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{DD.path <- system.file("extdata", "3b_SSM_DD_Example1.xlsx", package = "dbGaPCheckup", mustWork=TRUE)
DD.dict.A <- readxl::read_xlsx(DD.path)
path <- system.file("extdata", "DS_Example.txt", package = "dbGaPCheckup", mustWork=TRUE)
DS.data.A <- read.table(DS.path, header=TRUE, sep="\\t", quote="", as.is = TRUE)
save(DD.dict.A, DS.data.A, file = "ExampleA.rda")
}\if{html}{\out{</div>}}
}
\usage{
data(ExampleA)
}
\description{
Example data set and data dictionary with no errors.
}
