% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleB.R
\docType{data}
\name{ExampleB}
\alias{ExampleB}
\title{ExampleB}
\format{
\code{R data file} that contains two objects:
\describe{
\item{DD.dict.B}{Data dictionary}
\item{DS.data.B}{Data set}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{DD.path <- system.file("extdata", "3b_SSM_DD_Example1b.xlsx", package = "dbGaPCheckup", mustWork=TRUE)
DD.dict.B <- readxl::read_xlsx(DD.path)
DS.path <- system.file("extdata", "DS_Example1b.txt", package = "dbGaPCheckup", mustWork=TRUE)
DS.data.B <- read.table(DS.path, header=TRUE, sep="\\t", quote="", as.is = TRUE)
save(DD.dict.B, DS.data.B, file = "ExampleB.rda")
}\if{html}{\out{</div>}}
}
\usage{
data(ExampleB)
}
\description{
Example data set and data dictionary with intentional errors.
}
