% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleD.R
\docType{data}
\name{ExampleD}
\alias{ExampleD}
\title{ExampleD}
\format{
\code{R data file} that contains two objects:
\describe{
\item{DD.dict.D}{Data dictionary}
\item{DS.data.D}{Data set}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{path <- system.file("extdata", "3b_SSM_DD_Example2f.xlsx", package = "dbGaPCheckup", mustWork=TRUE)
DD.dict.D <- readxl::read_xlsx(path)
DS.path <- system.file("extdata", "DS_Example.txt", package = "dbGaPCheckup", mustWork=TRUE)
DS.data.D <- read.table(DS.path, header=TRUE, sep="\\t", quote="", as.is = TRUE)
save(DD.dict.D, DS.data.D, file = "ExampleD.rda")
}\if{html}{\out{</div>}}
}
\usage{
data(ExampleD)
}
\description{
Example data set and data dictionary with intentional errors.
}
