% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleF.R
\docType{data}
\name{ExampleF}
\alias{ExampleF}
\title{ExampleF}
\format{
\code{R data file} that contains two objects:
\describe{
\item{DD.dict.F}{Data dictionary}
\item{DS.data.F}{Data set}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{DD.path <- system.file("extdata", "3b_SSM_DD_Example4.xlsx", package = "dbGaPCheckup", mustWork=TRUE)
DD.dict.F <- readxl::read_xlsx(DD.path)
DS.path <- system.file("extdata", "DS_Example3d.txt", package = "dbGaPCheckup", mustWork=TRUE)
DS.data.F <- read.table(DS.path, header=TRUE, sep="\\t", quote="", as.is = TRUE)
save(DD.dict.F, DS.data.F, file = "ExampleF.rda")
}\if{html}{\out{</div>}}
}
\usage{
data(ExampleF)
}
\description{
Example data set and data dictionary with intentional errors.
}
