% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleL.R
\docType{data}
\name{ExampleL}
\alias{ExampleL}
\title{ExampleL}
\format{
\code{R data file} that contains two objects:
\describe{
\item{DD.dict.L}{Data dictionary}
\item{DS.data.L}{Data set}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{DD.path <- system.file("extdata", "3b_SSM_DD_Example2b.xlsx", package = "dbGaPCheckup", mustWork=TRUE)
DD.dict.L <- readxl::read_xlsx(DD.path)
DS.path <- system.file("extdata", "DS_Example2c.txt", package = "dbGaPCheckup", mustWork=TRUE)
DS.data.L <- read.table(DS.path, header=TRUE, sep="\\t", quote="", as.is = TRUE)
save(DD.dict.L, DS.data.L, file = "ExampleL.rda")
}\if{html}{\out{</div>}}
}
\usage{
data(ExampleL)
}
\description{
Example data set and data dictionary with intentional errors.
}
