% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleO.R
\docType{data}
\name{ExampleO}
\alias{ExampleO}
\title{ExampleO}
\format{
\code{R data file} that contains a single object:
\describe{
\item{DS.data.O}{Data set}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{DS.path <- system.file("extdata", "DS_Example3.txt", package = "dbGaPCheckup", mustWork=TRUE)
DS.data.O <- read.table(DS.path, header=TRUE, sep="\\t", quote="", as.is = TRUE)
save(DS.data.O, file = "ExampleO.rda")
}\if{html}{\out{</div>}}
}
\usage{
data(ExampleO)
}
\description{
Example data set with intentional errors.
}
