% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleR.R
\docType{data}
\name{ExampleR}
\alias{ExampleR}
\title{ExampleR}
\format{
\code{R data file} that contains two objects:
\describe{
\item{DD.dict.R}{Data dictionary}
\item{DS.data.R}{Data set}
}
}
\source{
\if{html}{\out{<div class="sourceCode">}}\preformatted{library(tidyverse)
DD.dict.R <- DD.dict.A
DS.data.R <- DS.data.A
# Change SUBJECT_ID to a string
DS.data.R$SUBJECT_ID <- paste0("A",DS.data.R$SUBJECT_ID)
DD.dict.R$TYPE[DD.dict.R$VARNAME=="SUBJECT_ID"] <- "string"
# Change HX_DEPRESSION to a string
DS.data.R <- DS.data.R \%>\% mutate(HX_DEPRESSION = recode(HX_DEPRESSION, '0' = 'no','1'='yes','-9999' = '-9999'))
DD.dict.R$TYPE[DD.dict.R$VARNAME=="HX_DEPRESSION"] <- "string"
DD.dict.R$VALUES[DD.dict.R$VARNAME=="HX_DEPRESSION"] <- "-9999=missing value"
# Set the extra VALUES column names to blank
DD.dict.R$`...18`[DD.dict.R$VARNAME=="HX_DEPRESSION"] <- NA
DD.dict.R$`...19`[DD.dict.R$VARNAME=="HX_DEPRESSION"] <- NA
nval <- which(names(DD.dict.R) == "VALUES")
names(DD.dict.R)[(nval + 1):ncol(DD.dict.R)] <- ""
save(DD.dict.R, DS.data.R, file="ExampleR.rda")
}\if{html}{\out{</div>}}
}
\usage{
data(ExampleR)
}
\description{
Example data set and data dictionary with no errors.
}
