% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleT.R
\docType{data}
\name{ExampleT}
\alias{ExampleT}
\title{ExampleT}
\format{
\code{R data file} that contains two objects:
\describe{
\item{DD.dict.T}{Data dictionary}
\item{DS.data.T}{Data set}
}
}
\usage{
data(ExampleT)
}
\description{
Example data set and data dictionary with intentional errors.
}
\keyword{datasets}
