% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_first_data.R
\name{id_first_data}
\alias{id_first_data}
\title{Relocate SUBJECT_ID to First Column of Data Set}
\usage{
id_first_data(DS.data)
}
\arguments{
\item{DS.data}{Data set.}
}
\value{
Updated data set with SUBJECT_ID as first column.
}
\description{
This utility function reorders the data set so that SUBJECT_ID comes first.
}
\details{
SUBJECT_ID is required to be the first column of the data set and first variable listed in the data dictionary.
}
\examples{
data(ExampleQ)
head(DS.data.Q)
DS.data.updated <- id_first_data(DS.data.Q)
head(DS.data.updated)
}
