% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/id_first_dict.R
\name{id_first_dict}
\alias{id_first_dict}
\title{Relocate SUBJECT_ID to First Column of Data Dictionary}
\usage{
id_first_dict(DD.dict)
}
\arguments{
\item{DD.dict}{Data dictionary.}
}
\value{
Updated data dictionary with SUBJECT_ID as first variable.
}
\description{
This utility function reorders the data dictionary so that SUBJECT_ID comes first.
}
\details{
SUBJECT_ID is required to be the first column of the data set and first variable listed in the data dictionary.
}
\examples{
data(ExampleQ)
head(DD.dict.Q)
DD.dict.updated <- id_first_dict(DD.dict.Q)
head(DD.dict.updated)
}
