% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NA_precheck.R
\name{NA_precheck}
\alias{NA_precheck}
\title{Min Max Required Pre-checks}
\usage{
NA_precheck(dict, data)
}
\arguments{
\item{dict}{Data dictionary.}

\item{data}{Data set.}
}
\value{
Tibble containing the following information for each check: (1) Time (time stamp); (2) Name (name of the function); (3) Status (Passed/Failed); (4) Message (A copy of the message the function printed out); (5) Information (More detailed information about the potential errors identified).
}
\description{
This function runs a workflow of the minimum number of checks required for a user to run minmax_check; the checks include \code{pkg_field_check}, \code{dimension_check}, and \code{name_check}.
}
\examples{
data(ExampleB)
NA_precheck(DD.dict.B, DS.data.B)
}
