% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_data.R
\name{reorder_data}
\alias{reorder_data}
\title{Reorder Data Set Utility Function}
\usage{
reorder_data(DD.dict, DS.data)
}
\arguments{
\item{DD.dict}{Data dictionary.}

\item{DS.data}{Data set.}
}
\value{
Updated data set with variables reordered to match the data dictionary.
}
\description{
This utility function reorders the data set to match the data dictionary.
}
\examples{
data(ExampleN)
name_check(DD.dict.N, DS.data.N)
DS.data_updated <- reorder_data(DD.dict.N, DS.data.N)
name_check(DD.dict.N, DS.data_updated)
}
