% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_dictionary.R
\name{reorder_dictionary}
\alias{reorder_dictionary}
\title{Reorder Data Dictionary Utility Function}
\usage{
reorder_dictionary(DD.dict, DS.data)
}
\arguments{
\item{DD.dict}{Data dictionary.}

\item{DS.data}{Data set.}
}
\value{
Updated data dictionary with variables reordered to match the data set.
}
\description{
This utility function reorders the data dictionary to match the data set.
}
\examples{
data(ExampleN)
name_check(DD.dict.N, DS.data.N)
DD.dict_updated <- reorder_dictionary(DD.dict.N, DS.data.N)
name_check(DD.dict_updated, DS.data.N)
}
