% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/super_short_precheck.R
\name{super_short_precheck}
\alias{super_short_precheck}
\title{Very Truncated Pre-check}
\usage{
super_short_precheck(dict, data)
}
\arguments{
\item{dict}{Data dictionary.}

\item{data}{Data set.}
}
\value{
Tibble containing the following information for each check: (1) Time (time stamp); (2) Name (name of the function); (3) Status (Passed/Failed); (4) Message (A copy of the message the function printed out); (5) Information (More detailed information about the potential errors identified).
}
\description{
This function runs a workflow of the minimum number of checks required for a user to run dbGaPCheckup_required_field_check; the checks include \code{dbGaP_required_field_check}, \code{dimension_check}, and \code{name_check}.
}
\examples{
# Example 1: Pass check
data(ExampleB)
super_short_precheck(DD.dict.B, DS.data.B)
}
