% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_meaning_table.R
\name{value_meaning_table}
\alias{value_meaning_table}
\title{Value-Meaning Table}
\usage{
value_meaning_table(DD.dict)
}
\arguments{
\item{DD.dict}{Data dictionary.}
}
\value{
A data frame with the columns VARNAME, TYPE, VALUE, MEANING.
}
\description{
This function generates a value-meaning table by parsing the VALUES fields.
}
\examples{
data(ExampleB)
head(value_meaning_table(DD.dict.B))
}
