% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_precheck.R
\name{values_precheck}
\alias{values_precheck}
\title{Values Pre-Check}
\usage{
values_precheck(dict)
}
\arguments{
\item{dict}{Data dictionary.}
}
\value{
Tibble containing the following information for each check: (1) Time (time stamp); (2) Name (name of the function); (3) Status (Passed/Failed); (4) Message (A copy of the message the function printed out); (5) Information (More detailed information about the potential errors identified).
}
\description{
This function runs a workflow of the minimum number of checks required for a user to run values_check; the checks include \code{field_check}, and \code{type_check}.
}
\examples{
data(ExampleB)
values_precheck(DD.dict.B)
}
