% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbGaPCheckup-package.R
\docType{package}
\name{dbGaPCheckup-package}
\alias{dbGaPCheckup}
\alias{dbGaPCheckup-package}
\title{dbGaPCheckup: dbGaP Checkup}
\description{
Contains functions that check for formatting of the Subject Phenotype data set and data dictionary as specified by the National Center for Biotechnology Information (NCBI) Database of Genotypes and Phenotypes (dbGaP) \url{https://www.ncbi.nlm.nih.gov/gap/docs/submissionguide/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://lwheinsberg.github.io/dbGaPCheckup/}
  \item \url{https://github.com/lwheinsberg/dbGaPCheckup}
  \item Report bugs at \url{https://github.com/lwheinsberg/dbGaPCheckup/issues}
}

}
\author{
\strong{Maintainer}: Lacey W. Heinsberg \email{law145@pitt.edu}

Authors:
\itemize{
  \item Daniel E. Weeks \email{weeks@pitt.edu}
}

Other contributors:
\itemize{
  \item  University of Pittsburgh [copyright holder]
}

}
\keyword{internal}
