% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicated_id_check.R
\name{duplicated_id_check}
\alias{duplicated_id_check}
\title{Duplicated ID Check}
\usage{
duplicated_id_check(DS.data, verbose = TRUE)
}
\arguments{
\item{DS.data}{Data set.}

\item{verbose}{When TRUE, the function prints the Message out, as well as more detailed diagnostic information.}
}
\value{
Tibble, returned invisibly, containing: (1) Time (Time stamp); (2) Name (Name of the function); (3) Status (Passed/Warning); (4) Message (A copy of the message the function printed out); (5) Information (Details about duplicated SUBJECT_ID values).
}
\description{
This function checks whether any duplicated SUBJECT_ID values are present in the dataset. While duplication may be expected in longitudinal data, it may indicate an error in cross-sectional submissions.
}
\details{
Duplicated Subject IDs are allowed in longitudinal data sets. This is an informational check.
}
\examples{
# Example 1: Warning, duplicated 'SUBJECT_ID'
data(ExampleT)
duplicated_id_check(DS.data.T)

# Example 2: Pass check 
data(ExampleA)
duplicated_id_check(DS.data.A)
}
\seealso{
\code{\link{id_check}}
}
