% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_check.R
\name{int_check}
\alias{int_check}
\title{Integer Check Base Function}
\usage{
int_check(data)
}
\arguments{
\item{data}{Number or vector of numbers.}
}
\value{
Logical, TRUE if all non-missing entries in the input vector are integers.
}
\description{
This function checks for integer values.
}
