% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_correct.R
\name{name_correct}
\alias{name_correct}
\title{Name Correction Utility Function}
\usage{
name_correct(DD.dict, DS.data)
}
\arguments{
\item{DD.dict}{Data dictionary.}

\item{DS.data}{Data set.}
}
\value{
Updated data set with variables renamed to match the data dictionary.
}
\description{
This utility function updates the data set so variable names match those listed in the data dictionary.
}
\details{
Recommend use with caution; perform name_check first.
}
\examples{
data(ExampleM)
name_check(DD.dict.M, DS.data.M)
DS.data_updated <- name_correct(DD.dict.M, DS.data.M)
name_check(DD.dict.M, DS.data_updated)
}
